efficient-vit copy/requirements.txt
absl-py==2.1.0
albucore==0.0.21
albumentations==1.4.23
altair==5.5.0
annotated-types==0.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
efficientnet_pytorch==0.7.1
einops==0.8.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
facenet-pytorch==2.6.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.0
moviepy==2.1.1
mpmath==1.3.0
mtcnn==1.0.0
narwhals==1.19.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==10.2.0
proglog==0.1.10
progress==1.6
protobuf==4.25.5
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentencepiece==0.2.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.1
sounddevice==0.5.1
soupsieve==2.6
streamlit==1.41.1
stringzilla==3.11.1
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.2.2
torchsummary==1.5.1
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vit-pytorch==1.8.9
watchdog==6.0.0
